+++ CHANGELOG.md
@@ -25,6 +25,17 @@
 * Add middleware for Express.js
 * Add new websocket tunneling service that can be used for local webhook development and testing
 
+### 6.5.1 / 2022-09-16
+* Add additional `Event` fields
+* Fix issue with `EventParticipant` not sending status on new event creation
+
+### 6.5.0 / 2022-07-29
+* Add `metadata` field to `JobStatus`
+* Add `interval_minutes` field in Scheduler booking config
+* Fixed `Event.originalStartTime` type to be `Date`
+* Fixed `SchedulerBookingOpeningHours.accountId` deserialization
+* Fixed json value for `confirmationEmailToHost` in `SchedulerBooking`
+
 ### 6.4.2 / 2022-06-14
 * Add `Message.save()` functionality for updating existing messages
 * Add missing `reminderMinutes` field in `Event`